move patch handling to eval pkg
